Fix #22 - change distances to ground to double

+++ cliff_detector/include/cliff_detector/cliff_detector.hpp
@@ -207,7 +207,7 @@ private:
   /// Class for managing sensor_msgs/CameraInfo messages
   image_geometry::PinholeCameraModel camera_model_;
   /// Calculated distances to ground for every row of depth image in mm
-  std::vector<unsigned> dist_to_ground_;
+  std::vector<double> dist_to_ground_;
   std::vector<double> tilt_compensation_factor_;
   std::vector<double> delta_row_;
